* Added option for double precision hillshading (2 triangle vs. 1 square cell render) * Switched elevation datatype from float to double across the board * Optimizations in TileRenderer * Rebuilt ModelOptions to be easier to extend

+++ jdem846/src-ui/us/wthr/jdem846/ui/OutputImageViewPanel.java
@@ -492,7 +492,7 @@ public class OutputImageViewPanel extends JdemPanel
 		String strMouseLatitude = formatter.format(mouseLatitude);
 		String strMouseLongitude = formatter.format(mouseLongitude);;
 		
-		float elevation = 0;
+		double elevation = 0;
 		try {
 			elevation = dataPackage.getElevation(trueY, trueX);
 		} catch (DataSourceException ex) {
