+++ jdem846/src/us/wthr/jdem846/input/edef/ElevationDatasetExchangeDataCache.java
@@ -69,9 +69,9 @@ public class ElevationDatasetExchangeDataCache extends DataCache
 		isDisposed = true;
 	}
 	
-	public float get(int position)
+	public double get(int position)
 	{
-		float elevation = 0;
+		double elevation = 0;
 		
 		int offset = (position * 4);
